Parse a FACSDiva Workspace

Function to parse a FACSDiva Workspace, generate a GatingHierarchy or GatingSet object, and associated flowCore gates.

diva_to_gatingset(
  obj,
  name = NULL,
  subset = NULL,
  path = obj@path,
  worksheet = c("normal", "global"),
  swap_cols = list(`FSC-H` = "FSC-W", `SSC-H` = "SSC-W"),
  verbose = FALSE,
  ...
)

Arguments

obj
diva_workspace

name
sample group to be parsed, either numeric index or the group name

subset
samples to be imported. either numeric index or the sample name. Default is NULL, which imports all samples.

path
the FCS data path

worksheet
select worksheet to import. either "normal" or "global"

swap_cols
diva seems to swap some data cols during importing fcs to experiments this argument provide a list to tell the parser which cols to be swapped default is list(`FSC-H` = 'FSC-W',`SSC-H` = 'SSC-W')

verbose
whether print more messages during the parsing

...
other arguments